feat: 修改

+++ services/charging_service/models.py
@@ -9,7 +9,7 @@ class ChargingRecord(Base):
  
     user_id = Column(String(50), nullable=False) 
     station_id = Column(Integer, nullable=False)
-    kwh_consumed = Column(Integer, nullable=False)
+    kwh_consumed = Column(Float, nullable=False)
     duration_minutes = Column(Float, nullable=False)
     created_at = Column(DateTime, server_default=func.now())
